+++ site/locales/en-US.js
@@ -47,7 +47,8 @@ module.exports = {
     timeline: 'Timeline',
     transfer: 'Transfer',
     switch: 'Switch',
-    rate: 'Rate'
+    rate: 'Rate',
+    preview: 'Preview'
   },
   designs: {
     'design-patterns': '设计模式',