+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 更新日志
 ## 2.14.1
 - 优化 `<Tree />` onBeforeSave 、 onDragStart 支持返回 Promise [#1179](https://github.com/XiaoMi/hiui/issues/1179)
+- 优化 `<DatePicker />` 新增属性 inputReadOnly 设置输入框为只读形态 [#1203](https://github.com/XiaoMi/hiui/issues/1203)
+- 优化 `<TimePicker />` 新增属性 inputReadOnly 设置输入框为只读形态 [#1203](https://github.com/XiaoMi/hiui/issues/1203)
 - 优化 `<TimePicker />` 新增hourStep、minuteStep、secondStep等时分秒间隔选项 [#1194](https://github.com/XiaoMi/hiui/issues/1195)
 - 修复 `<Carousel />`  Children 节点添加点击事件后，点击回调 index 异常问题 [#1201](https://github.com/XiaoMi/hiui/issues/1201)
 - 修复 `<Input />` ref 属性无效问题 [#1192](https://github.com/XiaoMi/hiui/issues/1192)